statistics bug fix

+++ app/src/main/java/com/ertools/demooder/presentation/components/widgets/EvaluationWidget.kt
@@ -77,7 +77,7 @@ fun EvaluationWidget(
                         progressAnimation.animateTo(
                             targetValue = 1f,
                             animationSpec = tween(
-                                durationMillis = (1000 * detectionPeriodSeconds.value.toDouble()).roundToInt(),
+                                durationMillis = (1000 * detectionPeriodSeconds.value).roundToInt(),
                                 easing = LinearEasing
                             )
                         )
